Wrote function spec for the code
I wrote the parameters for the Monte Carlo program.  These will need to be
expanded to accomodate the necessary statistical data.

An earlier draft of the code is still present in the bulk of the program.
This will be incorporated into various functions (i.e., I'll use good coding
practice, I promise).

Added a brief README file and the MIT license.